usm: migrate tests
